Updated winregsitry lib version
I tested with version 1.0.0 and the latest version 1.1.1 and neither worked. The truerng_test.py script is throwing an exception on line 509 due to the way the newer versions of the lib expect you to access the returned keys. Rolling back to 0.8.3 fixes the issue so I've updated this library install bat to download that specific version.